Feat: 노래 난이도 평가 API

+++ server/src/main/java/com/singsong/api/response/SongDetailRes.java
@@ -26,7 +26,7 @@ public class SongDetailRes {
     int songLikeCount;
     boolean isLiked;
     int songLevel;
-    int songEvalCount;
+    Long songEvalCount;
     int myEval;
     String songLyrics;
